third_party/blink/web_tests/external/wpt/webrtc-encoded-transform/script-transform-generateKeyFrame-simulcast.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: Cannot read properties of undefined (reading 'getReader')
[FAIL] generateKeyFrame works with simulcast rids
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame for rid that was negotiated away fails
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame with rid after simulcast->unicast negotiation fails
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
Harness: the test ran to completion.
